fix: move root snmp_server.py to project folder
