+++ pytest_snmpserver/snmp_server.py
@@ -4,6 +4,7 @@
 Simple SNMP server in pure Python
 """
 
+from __future__ import print_function
 
 import argparse
 import fnmatch
@@ -13,11 +14,13 @@ import socket
 import string
 import struct
 import sys
-import threading
-import time
 import types
+from contextlib import closing
 
-from collections import Iterable
+try:
+    from collections import Iterable
+except ImportError:
+    from collections.abc import Iterable
 
 try:
     from StringIO import StringIO
@@ -58,7 +61,11 @@ ASN1_GET_REQUEST_PDU = 0xA0
 ASN1_GET_NEXT_REQUEST_PDU = 0xA1
 ASN1_GET_RESPONSE_PDU = 0xA2
 ASN1_SET_REQUEST_PDU = 0xA3
+ASN1_TRAP_REQUEST_PDU = 0xA4
 ASN1_GET_BULK_REQUEST_PDU = 0xA5
+ASN1_INFORM_REQUEST_PDU = 0xA6
+ASN1_SNMPv2_TRAP_REQUEST_PDU = 0xA7
+ASN1_REPORT_REQUEST_PDU = 0xA8
 
 # error statuses
 ASN1_ERROR_STATUS_NO_ERROR = 0x00
@@ -106,19 +113,19 @@ SNMP_PDUS = (
 
 
 class ProtocolError(Exception):
-    """Raise when SNMP protocol error occured"""
+    """Raise when SNMP protocol error occurred"""
 
 
 class ConfigError(Exception):
-    """Raise when config error occured"""
+    """Raise when config error occurred"""
 
 
 class BadValueError(Exception):
-    """Raise when bad value error occured"""
+    """Raise when bad value error occurred"""
 
 
 class WrongValueError(Exception):
-    """Raise when wrong value (e.g. value not in available range) error occured"""
+    """Raise when wrong value (e.g. value not in available range) error occurred"""
 
 
 def encode_to_7bit(value):
@@ -162,16 +169,15 @@ def bytes_to_oid(data):
     while values:
         val = values.pop(0)
         if val > 0x7f:
-            huges = []
-            huges.append(val)
+            huge_vals = [val]
             while True:
                 next_val = values.pop(0)
-                huges.append(next_val)
+                huge_vals.append(next_val)
                 if next_val < 0x80:
                     break
             huge = 0
-            for i, huge_byte in enumerate(huges):
-                huge += (huge_byte & 0x7f) << (7 * (len(huges) - i - 1))
+            for i, huge_byte in enumerate(huge_vals):
+                huge += (huge_byte & 0x7f) << (7 * (len(huge_vals) - i - 1))
             res.append(huge)
         else:
             res.append(val)
@@ -221,11 +227,12 @@ def _read_int_len(stream, length, signed=False):
         result = twos_complement(result, 8 * length)
     return result
 
-
 def _write_int(value, strip_leading_zeros=True):
-    """Write int"""
+    """Write int while ensuring correct sign representation."""
     if abs(value) > 0xffffffffffffffff:
         raise Exception('Int value must be in [0..18446744073709551615]')
+    
+    # Determine the correct format specifier based on the value's magnitude and sign.
     if value < 0:
         if abs(value) <= 0x7f:
             result = struct.pack('>b', value)
@@ -235,12 +242,35 @@ def _write_int(value, strip_leading_zeros=True):
             result = struct.pack('>i', value)
         elif abs(value) <= 0x7fffffffffffffff:
             result = struct.pack('>q', value)
+        else:
+            raise Exception('Min signed int value')
     else:
-        result = struct.pack('>Q', value)
-    # strip first null bytes, if all are null - leave one
-    result = result.lstrip(b'\x00') if strip_leading_zeros else result
-    return result or b'\x00'
+        if not strip_leading_zeros:
+            # Always pack as the largest size to simplify leading zero handling.
+            if value <= 0x7fffffff:
+                result = struct.pack('>I', value)
+            else:
+                result = struct.pack('>Q', value)
+        else:
+            # Always pack as the largest size to simplify leading zero handling.
+            result = struct.pack('>Q', value)
+            # Check if the first relevant byte (ignoring leading zeros for now) would be misinterpreted as negative.
+            if (result[0] == 0x00 and (result[1] & 0x80) != 0):
+                # If not stripping leading zeros, or if stripping them would cause a misinterpretation,
+                # leave the result as is. This branch might need revisiting based on specific needs.
+                pass
+            else:
+                # Here's the core of the adjustment: only strip leading zeros if it does not lead to misinterpretation.
+                # This means checking if the first byte would make it look negative and adjusting accordingly.
+                first_non_zero_byte = next((i for i, byte in enumerate(result) if byte != 0), len(result) - 1)
+                if result[first_non_zero_byte] & 0x80:
+                    # If the first non-zero byte's MSB is set, prepend a 0x00 to keep it positive.
+                    result = b'\x00' + result[first_non_zero_byte:]
+                else:
+                    # Otherwise, strip all leading zeros except the last one, if all are zeros.
+                    result = result[first_non_zero_byte:]
 
+    return result or b'\x00'
 
 def _write_asn1_length(length):
     """Write ASN.1 length"""
@@ -276,7 +306,7 @@ def _parse_asn1_octet_string(stream):
     length = _parse_asn1_length(stream)
     value = stream.read(length)
     # if any char is not printable - convert string to hex
-    if any([c not in string.printable for c in value]):
+    if any(c not in string.printable for c in value):
         return ' '.join(['%02X' % ord(x) for x in value])
     return value
 
@@ -343,10 +373,44 @@ def _parse_asn1_opaque(stream):
     return stream.read(length)
 
 
+def _is_trap_request(result):
+    """Checks if it is Trap-PDU request."""
+    return len(result) > 2 and result[2][1] == ASN1_TRAP_REQUEST_PDU
+
+
+def _validate_protocol(pdu_index, tag, result):
+    """Validates the protocol and returns True if valid, or False otherwise."""
+    if _is_trap_request(result):
+        if (
+                pdu_index == 4 and tag != ASN1_OBJECT_IDENTIFIER or
+                pdu_index == 5 and tag != ASN1_IPADDRESS or
+                pdu_index in [6, 7] and tag != ASN1_INTEGER or
+                pdu_index == 8 and tag != ASN1_TIMETICKS
+        ):
+            return False
+    elif (
+            pdu_index in [1, 4, 5, 6] and tag != ASN1_INTEGER or
+            pdu_index == 2 and tag != ASN1_OCTET_STRING or
+            pdu_index == 3 and tag not in [
+                ASN1_GET_REQUEST_PDU,
+                ASN1_GET_NEXT_REQUEST_PDU,
+                ASN1_SET_REQUEST_PDU,
+                ASN1_GET_BULK_REQUEST_PDU,
+                ASN1_TRAP_REQUEST_PDU,
+                ASN1_INFORM_REQUEST_PDU,
+                ASN1_SNMPv2_TRAP_REQUEST_PDU,
+            ]
+    ):
+        return False
+    return True
+
+
 def _parse_snmp_asn1(stream):
     """Parse SNMP ASN.1
     After |IP|UDP| headers and "sequence" tag, SNMP protocol data units (PDUs) are the next:
     |version|community|PDU-type|request-id|error-status|error-index|variable bindings|
+    but for ASN1_TRAP_REQUEST_PDU next:
+    |version|community|PDU-type|enterprise-oid|agent|trap-type|specific-type|uptime|
     """
     result = []
     wait_oid_value = False
@@ -359,16 +423,7 @@ def _parse_snmp_asn1(stream):
             return result
         tag = ord(read_byte)
         # check protocol's tags at indices
-        if (
-                pdu_index in [1, 4, 5, 6] and tag != ASN1_INTEGER or
-                pdu_index == 2 and tag != ASN1_OCTET_STRING or
-                pdu_index == 3 and tag not in [
-                    ASN1_GET_REQUEST_PDU,
-                    ASN1_GET_NEXT_REQUEST_PDU,
-                    ASN1_SET_REQUEST_PDU,
-                    ASN1_GET_BULK_REQUEST_PDU,
-                ]
-        ):
+        if not _validate_protocol(pdu_index, tag, result):
             raise ProtocolError('Invalid tag for PDU unit "{}"'.format(SNMP_PDUS[pdu_index]))
         if tag == ASN1_SEQUENCE:
             length = _parse_asn1_length(stream)
@@ -378,7 +433,7 @@ def _parse_snmp_asn1(stream):
             value = _read_int_len(stream, length, True)
             logger.debug('ASN1_INTEGER: %s', value)
             # pdu_index is version, request-id, error-status, error-index
-            if wait_oid_value or pdu_index in [1, 4, 5, 6]:
+            if wait_oid_value or pdu_index in [1, 4, 5, 6] or _is_trap_request(result):
                 result.append(('INTEGER', value))
                 wait_oid_value = False
         elif tag == ASN1_OCTET_STRING:
@@ -420,18 +475,39 @@ def _parse_snmp_asn1(stream):
             logger.debug('ASN1_SET_REQUEST_PDU: %s', 'length = {}'.format(length))
             if pdu_index == 3:  # PDU-type
                 result.append(('ASN1_SET_REQUEST_PDU', tag))
+        elif tag == ASN1_TRAP_REQUEST_PDU:
+            length = _parse_asn1_length(stream)
+            logger.debug('ASN1_TRAP_REQUEST_PDU: %s', 'length = {}'.format(length))
+            if pdu_index == 3:  # PDU-type
+                result.append(('ASN1_TRAP_REQUEST_PDU', tag))
+        elif tag == ASN1_INFORM_REQUEST_PDU:
+            if result and result[0][1] == 0:
+                raise Exception('INFORM request PDU is not supported in SNMPv1!')
+            length = _parse_asn1_length(stream)
+            logger.debug('ASN1_INFORM_REQUEST_PDU: %s', 'length = {}'.format(length))
+            if pdu_index == 3:  # PDU-type
+                result.append(('ASN1_INFORM_REQUEST_PDU', tag))
+        elif tag == ASN1_SNMPv2_TRAP_REQUEST_PDU:
+            if result and result[0][1] == 0:
+                raise Exception('SNMPv2 TRAP PDU request is not supported in SNMPv1!')
+            length = _parse_asn1_length(stream)
+            logger.debug('ASN1_SNMPv2_TRAP_REQUEST_PDU: %s', 'length = {}'.format(length))
+            if pdu_index == 3:  # PDU-type
+                result.append(('ASN1_SNMPv2_TRAP_REQUEST_PDU', tag))
+        elif tag == ASN1_REPORT_REQUEST_PDU:
+            raise Exception('Report request PDU is not supported!')
         elif tag == ASN1_TIMETICKS:
             length = _read_byte(stream)
             value = _read_int_len(stream, length)
             logger.debug('ASN1_TIMETICKS: %s (%s)', value, timeticks_to_str(value))
-            if wait_oid_value:
+            if wait_oid_value or _is_trap_request(result):
                 result.append(('TIMETICKS', value))
                 wait_oid_value = False
         elif tag == ASN1_IPADDRESS:
             length = _read_byte(stream)
             value = _read_int_len(stream, length)
             logger.debug('ASN1_IPADDRESS: %s (%s)', value, int_to_ip(value))
-            if wait_oid_value:
+            if wait_oid_value or _is_trap_request(result):
                 result.append(('IPADDRESS', int_to_ip(value)))
                 wait_oid_value = False
         elif tag == ASN1_COUNTER32:
@@ -475,11 +551,10 @@ def _parse_snmp_asn1(stream):
         elif tag == ASN1_END_OF_MIB_VIEW:
             value = _read_byte(stream)
             logger.debug('ASN1_END_OF_MIB_VIEW: %s', value)
-            return (('', ''), ('', ''))
+            return ('', ''), ('', '')
         else:
             logger.debug('?: %s', hex(ord(read_byte)))
         pdu_index += 1
-    return result
 
 
 def get_next_oid(oid):
@@ -512,9 +587,8 @@ def boolean(value):
 
 def integer(value, enum=None):
     """Get Integer"""
-    if enum and isinstance(enum, Iterable):
-        if not value in enum:
-            raise WrongValueError('Integer value {} is outside the range of enum values'.format(value))
+    if enum and isinstance(enum, Iterable) and value not in enum:
+        raise WrongValueError('Integer value {} is outside the range of enum values'.format(value))
     if not (-2147483648 <= value <= 2147483647):
         raise Exception('Integer value must be in [-2147483648..2147483647]')
     if not enum:
@@ -587,7 +661,6 @@ def uint64(value):
         'BB', ASN1_OPAQUE_TAG1, ASN1_OPAQUE_UINT64
     ) + _write_asn1_length(len(uint64_value)) + uint64_value
     return write_tv(ASN1_OPAQUE, opaque_type_value)
-
 
 
 def utf8_string(value):
@@ -676,7 +749,6 @@ def get_next(oids, oid):
 
 def parse_config(filename):
     """Read and parse a config"""
-    oids = {}
     try:
         with open(filename, 'rb') as conf_file:
             data = conf_file.read()
@@ -684,9 +756,8 @@ def parse_config(filename):
             exec(data, globals(), out_locals)
             oids = out_locals['DATA']
             for value in oids.values():
-                if isinstance(value, types.FunctionType):
-                    if value.__code__.co_argcount != 1:
-                        raise ConfigError('"{}" must have one argument'.format(value.__name__))
+                if isinstance(value, types.FunctionType) and value.__code__.co_argcount != 1:
+                    raise ConfigError('"{}" must have one argument'.format(value.__name__))
             return oids
     except Exception as ex:
         raise ConfigError('Config parsing error: {}'.format(ex))
@@ -732,16 +803,14 @@ def handle_get_request(oids, oid):
     return error_status, error_index, oid_value
 
 
-def handle_get_next_request(oids, oid):
+def handle_get_next_request(oids, oid, limit_to_last_in_config=True):
     """Handle GetNextRequest"""
     error_status = ASN1_ERROR_STATUS_NO_ERROR
     error_index = 0
-    oid_value = null()
-    new_oid = None
     if oid in oids:
         new_oid = get_next(oids, oid)
         if not new_oid:
-            oid_value = struct.pack('BB', ASN1_END_OF_MIB_VIEW, 0)  #null()
+            oid_value = struct.pack('BB', ASN1_END_OF_MIB_VIEW, 0)
         else:
             oid_value = oids.get(new_oid)
     else:
@@ -766,6 +835,11 @@ def handle_get_next_request(oids, oid):
         oid = get_next_oid(oid.rstrip('.0')) + '.0'
     # if wildcards are used in oid - replace them
     final_oid = replace_wildcards(oid)
+    # to prevent loop - check a new oid and if it is more than the last in config - stop 
+    if oids and limit_to_last_in_config:
+        last_oid_in_config = sorted(oids, key=functools.cmp_to_key(oid_cmp))[-1]
+        if oid_cmp(final_oid, last_oid_in_config) > 0:
+            oid_value = struct.pack('BB', ASN1_END_OF_MIB_VIEW, 0)
     return error_status, error_index, final_oid, oid_value
 
 
@@ -774,39 +848,94 @@ def handle_set_request(oids, oid, type_and_value):
     error_status = ASN1_ERROR_STATUS_NO_ERROR
     error_index = 0
     value_type, value = type_and_value
+
+    res = None
     if value_type == 'INTEGER':
         enum_values = None
-        if isinstance(oids[oid], tuple) and len(oids[oid]) > 1:
+        if oid in oids and isinstance(oids[oid], tuple) and len(oids[oid]) > 1:
             enum_values = oids[oid][1]
-        oids[oid] = integer(value, enum=enum_values)
+        res = integer(value, enum=enum_values)
     elif value_type == 'STRING':
-        oids[oid] = octet_string(value if PY3 else value.encode('latin'))
+        res = octet_string(value if PY3 else value.encode('latin'))
     elif value_type == 'OID':
-        oids[oid] = object_identifier(value)
+        res = object_identifier(value)
     elif value_type == 'TIMETICKS':
-        oids[oid] = timeticks(value)
+        res = timeticks(value)
     elif value_type == 'IPADDRESS':
-        oids[oid] = ip_address(value)
+        res = ip_address(value)
     elif value_type == 'COUNTER32':
-        oids[oid] = counter32(value)
+        res = counter32(value)
     elif value_type == 'COUNTER64':
-        oids[oid] = counter64(value)
+        res = counter64(value)
     elif value_type == 'GAUGE32':
-        oids[oid] = gauge32(value)
+        res = gauge32(value)
     elif value_type == 'OPAQUE':
         if value[0] == 'FLOAT':
-            oids[oid] = real(value[1])
+            res = real(value[1])
         elif value[0] == 'DOUBLE':
-            oids[oid] = double(value[1])
+            res = double(value[1])
         elif value[0] == 'UINT64':
-            oids[oid] = uint64(value[1])
+            res = uint64(value[1])
         elif value[0] == 'INT64':
-            oids[oid] = int64(value[1])
+            res = int64(value[1])
         else:
             raise Exception('Unsupported type: {} ({})'.format(value_type, repr(value)))
-    oid_value = oids[oid]
+        
+    if res is None:
+        error_status = ASN1_ERROR_STATUS_WRONG_VALUE
+        error_index = 1
+        res = null()
+    else:
+        if oid in oids and isinstance(oids[oid], types.FunctionType):
+            oids[oid](oid, value)
+        else:
+            oids[oid] = res
+
+    oid_value = res
     return error_status, error_index, oid_value
 
+def handle_trap_request(oids, oid, type_and_value, details):
+    """Handle Trap Request"""
+    logger.info(type_and_value)
+    value_type, value = type_and_value
+
+    res = None
+    if value_type == 'INTEGER':
+        enum_values = None
+        if oid in oids and isinstance(oids[oid], tuple) and len(oids[oid]) > 1:
+            enum_values = oids[oid][1]
+        res = integer(value, enum=enum_values)
+    elif value_type == 'STRING':
+        res = octet_string(value if PY3 else value.encode('latin'))
+    elif value_type == 'OID':
+        res = object_identifier(value)
+    elif value_type == 'TIMETICKS':
+        res = timeticks(value)
+    elif value_type == 'IPADDRESS':
+        res = ip_address(value)
+    elif value_type == 'COUNTER32':
+        res = counter32(value)
+    elif value_type == 'COUNTER64':
+        res = counter64(value)
+    elif value_type == 'GAUGE32':
+        res = gauge32(value)
+    elif value_type == 'OPAQUE':
+        if value[0] == 'FLOAT':
+            res = real(value[1])
+        elif value[0] == 'DOUBLE':
+            res = double(value[1])
+        elif value[0] == 'UINT64':
+            res = uint64(value[1])
+        elif value[0] == 'INT64':
+            res = int64(value[1])
+        else:
+            raise Exception('Unsupported type: {} ({})'.format(value_type, repr(value)))
+        
+    if res is not None:
+        if oid in oids and isinstance(oids[oid], types.FunctionType):
+            oids[oid](oid, value, details)
+        else:
+            oids[oid] = res
 
 def craft_response(version, community, request_id, error_status, error_index, oid_items):
     """Craft SNMP response"""
@@ -842,176 +971,192 @@ def craft_response(version, community, request_id, error_status, error_index, oi
     return response
 
 
-def generate_response(request_result, oids):
+def snmp_server(host, port, oids):
+    """Main SNMP server loop"""
+    with closing(socket.socket(socket.AF_INET, socket.SOCK_DGRAM, socket.IPPROTO_UDP)) as sock:
+        sock.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
+        sock.bind((host, port))
+        print('SNMP server listening on {}:{}'.format(host, port))
 
-    # get required fields from request
-    version = request_result[0][1]
-    community = request_result[1][1]
-    pdu_type = request_result[2][1]
-    request_id = request_result[3][1]
-    max_repetitions = request_result[5][1]
-    logger.debug('max_repetitions %i', max_repetitions)
+        # SNMP server main loop
+        while True:
+            request_data, address = sock.recvfrom(4096)
+            logger.debug('Received %d bytes from %s', len(request_data), address)
 
-    error_status = ASN1_ERROR_STATUS_NO_ERROR
-    error_index = 0
-    oid_items = []
-    oid_value = null()
+            request_stream = StringIO(request_data.decode('latin'))
+            try:
+                request_result = _parse_snmp_asn1(request_stream)
+            except ProtocolError as ex:
+                logger.error('SNMP request parsing failed: %s', ex)
+                continue
 
-    # handle protocol data units
-    if pdu_type == ASN1_GET_REQUEST_PDU:
-        requested_oids = request_result[6:]
-        for _, oid in requested_oids:
-            _, _, oid_value = handle_get_request(oids, oid)
-            # if oid value is a function - call it to get the value
-            if isinstance(oid_value, types.FunctionType):
-                oid_value = oid_value(oid)
-            if isinstance(oid_value, tuple):
-                oid_value = oid_value[0]
-            oid_items.append((oid_to_bytes(oid), oid_value))
+            # get required fields from request
+            version = request_result[0][1]
+            community = request_result[1][1]
+            pdu_type = request_result[2][1]
+            request_id = request_result[3][1]
 
-    elif pdu_type == ASN1_GET_NEXT_REQUEST_PDU:
-        oid = request_result[6][1]
-        error_status, error_index, oid, oid_value = handle_get_next_request(oids, oid)
-        if isinstance(oid_value, types.FunctionType):
-            oid_value = oid_value(oid)
-        if isinstance(oid_value, tuple):
-            oid_value = oid_value[0]
-        oid_items.append((oid_to_bytes(oid), oid_value))
+            expected_length = 8 if pdu_type == ASN1_TRAP_REQUEST_PDU else 7
+            if len(request_result) < expected_length:
+                raise Exception('Invalid ASN.1 parsed request length! %s' % str(request_result))
 
-    elif pdu_type == ASN1_GET_BULK_REQUEST_PDU:
-        requested_oids = request_result[6:]
-        for _ in range(0, max_repetitions):
-            for idx, val in enumerate(requested_oids):
-                oid = val[1]
+            error_status = ASN1_ERROR_STATUS_NO_ERROR
+            error_index = 0
+            oid_items = []
+            oid_value = null()
+
+            # handle protocol data units
+            if pdu_type == ASN1_GET_REQUEST_PDU:
+                requested_oids = request_result[6:]
+                for _, oid in requested_oids:
+                    _, _, oid_value = handle_get_request(oids, oid)
+                    # if oid value is a function - call it to get the value
+                    if isinstance(oid_value, types.FunctionType):
+                        oid_value = oid_value(oid)
+                    if isinstance(oid_value, tuple):
+                        oid_value = oid_value[0]
+                    oid_items.append((oid_to_bytes(oid), oid_value))
+            elif pdu_type == ASN1_GET_NEXT_REQUEST_PDU:
+                oid = request_result[6][1]
                 error_status, error_index, oid, oid_value = handle_get_next_request(oids, oid)
                 if isinstance(oid_value, types.FunctionType):
                     oid_value = oid_value(oid)
                 if isinstance(oid_value, tuple):
                     oid_value = oid_value[0]
                 oid_items.append((oid_to_bytes(oid), oid_value))
-                requested_oids[idx] = ('OID', oid)
+            elif pdu_type == ASN1_GET_BULK_REQUEST_PDU:
+                max_repetitions = request_result[5][1]
+                logger.debug('max_repetitions: %i', max_repetitions)
+                requested_oids = request_result[6:]
+                for _ in range(0, max_repetitions):
+                    for idx, val in enumerate(requested_oids):
+                        oid = val[1]
+                        error_status, error_index, oid, oid_value = handle_get_next_request(oids, oid)
+                        if isinstance(oid_value, types.FunctionType):
+                            oid_value = oid_value(oid)
+                        if isinstance(oid_value, tuple):
+                            oid_value = oid_value[0]
+                        oid_items.append((oid_to_bytes(oid), oid_value))
+                        requested_oids[idx] = ('OID', oid)
+            elif pdu_type == ASN1_SET_REQUEST_PDU:
+                if len(request_result) < 8:
+                    raise Exception('Invalid ASN.1 parsed request length for SNMP set request!')
+                oid = request_result[6][1]
+                type_and_value = request_result[7]
+                try:
+                    if oid in oids and isinstance(oids[oid], tuple) and len(oids[oid]) > 1:
+                        enum_values = oids[oid][1]
+                        new_value = type_and_value[1]
+                        if isinstance(enum_values, Iterable) and new_value not in enum_values:
+                            raise WrongValueError('Value {} is outside the range of enum values'.format(new_value))
+                    logger.debug('oid: %s, type_and_value: %s', oid, type_and_value)
+                    error_status, error_index, oid_value = handle_set_request(oids, oid, type_and_value)
+                except WrongValueError as ex:
+                    logger.error(f"Wrong value error: {ex}")
+                    error_status = ASN1_ERROR_STATUS_WRONG_VALUE
+                    error_index = 0
+                except Exception as ex:
+                    logger.error(f"Unknown exception error: {ex}")
+                    error_status = ASN1_ERROR_STATUS_BAD_VALUE
+                    error_index = 0
+                # if oid value is a function - call it to get the value
+                if isinstance(oid_value, types.FunctionType):
+                    oid_value = oid_value(oid)
+                if isinstance(oid_value, tuple):
+                    oid_value = oid_value[0]
+                oid_items.append((oid_to_bytes(oid), oid_value))
+            elif pdu_type == ASN1_INFORM_REQUEST_PDU:
+                if len(request_result) < 8:
+                    raise Exception('Invalid ASN.1 parsed request length for SNMP set request!')
+                requested_oids = request_result[6:]
+                if len(requested_oids) % 2:
+                    raise Exception('Invalid length of OID and value items in SNMP inform request!')
+                for i in range(0, len(requested_oids), 2):
+                    oid = requested_oids[i][1]
+                    type_and_value = requested_oids[i + 1]
+                    error_status, error_index, oid_value = handle_set_request(oids, oid, type_and_value)
+                    if isinstance(oid_value, types.FunctionType):
+                        oid_value = oid_value(oid)
+                    if isinstance(oid_value, tuple):
+                        oid_value = oid_value[0]
+                    oid_items.append((oid_to_bytes(oid), oid_value))
+            elif pdu_type == ASN1_TRAP_REQUEST_PDU:
+                if len(request_result) < 10:
+                    raise Exception('Invalid ASN.1 parsed request length for SNMP trap request!')
+                logger.debug('Trap request length %d', len(request_result))
+                agent_community = request_result[1][1]
+                enterprise_oid = request_result[3][1]
+                agent_addr = request_result[4][1]
+                trap_type = request_result[5][1]
+                specific_type = request_result[6][1]
+                uptime = request_result[7][1]
+                oid = request_result[8][1]
+                type_and_value = request_result[9]
+                handle_trap_request(oids, oid, type_and_value, {
+                    'agent_community': agent_community,
+                    'enterprise_oid': enterprise_oid,
+                    'agent_addr': agent_addr,
+                    'trap_type': trap_type,
+                    'specific_type': specific_type,
+                    'uptime': uptime,
+                })
+                continue # traps do not require a response
+            else:
+                continue
 
-    elif pdu_type == ASN1_SET_REQUEST_PDU:
-        if len(request_result) < 8:
-            raise Exception('Invalid ASN.1 parsed request length for SNMP set request!')
-        oid = request_result[6][1]
-        type_and_value = request_result[7]
-        try:
-            if isinstance(oids[oid], tuple) and len(oids[oid]) > 1:
-                enum_values = oids[oid][1]
-                new_value = type_and_value[1]
-                if isinstance(enum_values, Iterable) and new_value not in enum_values:
-                    raise WrongValueError('Value {} is outside the range of enum values'.format(new_value))
-            error_status, error_index, oid_value = handle_set_request(oids, oid, type_and_value)
-        except WrongValueError as ex:
-            logger.error(ex)
-            error_status = ASN1_ERROR_STATUS_WRONG_VALUE
-            error_index = 0
-        except Exception as ex:
-            logger.error(ex)
-            error_status = ASN1_ERROR_STATUS_BAD_VALUE
-            error_index = 0
-        # if oid value is a function - call it to get the value
-        if isinstance(oid_value, types.FunctionType):
-            oid_value = oid_value(oid)
-        if isinstance(oid_value, tuple):
-            oid_value = oid_value[0]
-        oid_items.append((oid_to_bytes(oid), oid_value))
-
-    # craft SNMP response
-    response = craft_response(
-        version, community, request_id, error_status, error_index, oid_items)
-    return response
-
-def send_response(sock, response, address):
-    logger.debug('Sending %d bytes of response', len(response))
-    try:
-        sock.sendto(response, address)
-    except socket.error as ex:
-        logger.error('Failed to send %d bytes of response: %s', len(response), ex)
-    logger.debug('')
-
-
-class SNMPServer:
-    DEFAULT_LISTEN_HOST = '0.0.0.0'
-    DEFAULT_LISTEN_PORT = 0
-    DELAY_BEFORE_REPLY = 0
-
-    def __init__(self, host, port):
-        self.host = host
-        self.port = port
-        self.sock = socket.socket(socket.AF_INET, socket.SOCK_DGRAM, socket.IPPROTO_UDP)
-        self.sock.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
-        self.expected_messages = dict()
-        self._is_running = True
-
-    def start(self):
-        self.sock.bind((self.host, self.port))
-        if self.port == 0:
-            self.port = self.sock.getsockname()[1]
-        logger.info('SNMP server listening on {}:{}'.format(self.host, self.port))
-
-    def stop(self):
-        self._is_running = False
-        self.sock.close()
-
-    def __enter__(self):
-        self.start()
-        return self
-
-    def __exit__(self, exception_type, exception_vale, traceback):
-        self.stop()
-
-    def process_request(self):
-        while self._is_running:
-            request_data, address = self.sock.recvfrom(4096)
-            logger.debug('Received %d bytes from %s', len(request_data), address)
-
-            request_stream = StringIO(request_data.decode('latin'))
-            request_result = _parse_snmp_asn1(request_stream)
-
-            if len(request_result) < 7:
-                raise Exception('Invalid ASN.1 parsed request length!')
-            request = dict(request_result)
-            print(request)
-
-            _id = request['OID']
-            if _id not in self.expected_messages:
-                self.sock.close()
-                raise ValueError(f'Request OID ({_id}) was not expected')
-
-            response = generate_response(request_result, self.expected_messages)
-            time.sleep(self.DELAY_BEFORE_REPLY)
-            send_response(self.sock, response, address)
-
-    def expect_request(self, request_id, reply_with, populate_parent=True):
-        if isinstance(reply_with, str):
-            reply = octet_string(reply_with)
-        elif isinstance(reply_with, int):
-            reply = integer(reply_with)
-        elif isinstance(reply_with, list):
-            reply = integer(reply_with[0], enum=reply_with)
-        else:
-            print('wtf!!', reply_with)
-            reply = reply_with
-
-        print(request_id, reply)
-        self.expected_messages[request_id] = reply
-        if populate_parent:
-            parent = request_id.rpartition('.')[0]
-            if parent not in self.expected_messages:
-                self.expected_messages[parent] = None
-
+            # craft SNMP response
+            logger.info('Crafting response')
+            logger.info(request_id)
+            response = craft_response(
+                version, community, request_id, error_status, error_index, oid_items)
+            logger.debug('Sending %d bytes of response', len(response))
+            try:
+                sock.sendto(response, address)
+            except socket.error as ex:
+                logger.error('Failed to send %d bytes of response: %s', len(response), ex)
+            logger.debug('')
 
 def main():
+    """Main"""
+    parser = argparse.ArgumentParser(description='SNMP server')
+    parser.add_argument(
+        '-p', '--port', dest='port', type=int,
+        help='port (by default 161 - requires root privileges)', default=161, required=False)
+    parser.add_argument(
+        '-c', '--config', type=str,
+        help='OIDs config file', required=False)
+    parser.add_argument(
+        '-d', '--debug',
+        help='run in debug mode', action='store_true')
+    parser.add_argument(
+        '-v', '--version', action='version',
+        version='SNMP server v{}'.format(__version__))
+
+    args = parser.parse_args()
+
+    if args.debug:
+        logger.setLevel(logging.DEBUG)
+
+    # work as an echo server if no config is passed
+    oids = {
+        '*': lambda oid: octet_string(oid),
+    }
+    # read a config
+    config_filename = args.config
+    if config_filename:
+        try:
+            oids = parse_config(config_filename)
+        except ConfigError as ex:
+            logger.error(ex)
+            sys.exit(-1)
+
     host = '0.0.0.0'
-    port = 1234
-    s = SNMPServer(host, port)
-    s.start()
-    print(s.expect_request('1.3.6.1.2.1.2.2.1.2'))
-    s.stop()
+    port = args.port
+    try:
+        snmp_server(host, port, oids)
+    except KeyboardInterrupt:
+        logger.debug('Interrupted by Ctrl+C')
 
 
 if __name__ == '__main__':
     main()
-
